+++ libcaf_core/src/outbound_path.cpp
@@ -56,7 +56,7 @@ void outbound_path::emit_open(local_actor* self, stream_slot slot,
                                  nullptr, prio});
 }
 
-void outbound_path::emit_batch(local_actor* self, long xs_size, message xs) {
+void outbound_path::emit_batch(local_actor* self, int32_t xs_size, message xs) {
   CAF_LOG_TRACE(CAF_ARG(slots) << CAF_ARG(xs_size) << CAF_ARG(xs));
   CAF_ASSERT(xs_size > 0);
   CAF_ASSERT(xs_size <= std::numeric_limits<int32_t>::max());
